### 2019-09-16 d4d247f
fix: typo in bibliography file
