### 2019-09-16 38207e4
fix: updating author names
