+++ joss/paper.bib
@@ -62,7 +62,7 @@
     }
 
 @article{mudd2014statistical,
-    author = {Mudd, Simon M. and Attal, Mikaël and Milodowski, David T. and Grieve, Stuart W. D. and Valters, Declan A.},
+    author = {Mudd, S. M. and Attal, M. and Milodowski, D. T. and Grieve, S. W. D. and Valters, D. A.},
     title = {A statistical framework to quantify spatial variation in channel gradients using the integral method of channel profile analysis},
     journal = {Journal of Geophysical Research: Earth Surface},
     volume = {119},
@@ -89,6 +89,7 @@
 
 
 @article{schwanghart2010topo,
+    author = "Schwanghart, W. and Kuhn, N. J.",
     title = "TopoToolbox: A set of Matlab functions for topographic analysis",
     journal = "Environmental Modelling & Software",
     volume = "25",
