### 2019-09-16 a958931
fix: adding missing references
