+++ joss/paper.bib
@@ -100,3 +100,14 @@
     doi = "https://doi.org/10.1016/j.envsoft.2009.12.002",
     url = "http://www.sciencedirect.com/science/article/pii/S1364815209003053"
     }
+
+@Article{	  perron2013integral,
+    author	= {Perron, J Taylor and Royden, Leigh},
+    journal	= {Earth Surface Processes and Landforms},
+    number	= {6},
+    pages		= {570--576},
+    publisher	= {Wiley Online Library},
+    title		= {An integral approach to bedrock river profile analysis},
+    volume	= {38},
+    year		= {2013}
+  }
