+++ joss/paper.bib
@@ -22,18 +22,11 @@
     DOI = {10.5194/gmd-12-1267-2019}
     }
 
-@Article{adams2017dakotatheory,
-      address	= {Albuquerque, NM},
-      author	= {B.M. Adams and L.E. Bauman and W.J. Bohnhoff and K.R.
-    		  Dalbey and M.S. Ebeida and J.P. Eddy and M.S. Eldred and
-    		  P.D. Hough and K.T. Hu and J.D. Jakeman and J.A. Stephens
-    		  and L.P. Swiler and D.M. Vigil and T.M. Wildey},
-      journal	= {Sandia Technical Report SAND2014-4253},
-      title		= {Dakota, A Multilevel Parallel Object-Oriented Framework
-    		  for Design Optimization, Parameter Estimation, Uncertainty
-    		  Quantification, and Sensitivity Analysis: Version 6.6
-    		  Theory Manual},
-      year		= {2017}
+@article{adams2019dakota,
+    author = {Adams, Brian M and Eldred, Michael S and Geraci, Gianluca and Hooper, Russell W and Jakeman, John D and Maupin, Kathryn A and Monschk, Jason A and Rushdi, Ahmad A and Stephens, J Adam and Swiler, Laura P and Wildey, Timothy M and Bohnhoff, William J and Dalbey, Keith R and Ebeida, Mohamed S and Eddy, John P and Hough, Patricia D and Khalil, Mohammad and Hu, Kenneth T and Ridgway, Elliott M and Vigil, Dena M and Winokur, Justin G},
+    title = {{Dakota, A Multilevel Parallel Object-Oriented Framework for Design Optimization, Parameter Estimation, Uncertainty Quantification, and Sensitivity Analysis: Version 6.10 User{\textquoteright}s Manual}},
+    journal = {Sandia National Laboratories, Tech. Rep. SAND2014-4633},
+    year = {2019}
     }
 
 
