### 2019-10-01 #1776 @whedon
[REVIEW]: umami: A Python package for Earth surface dynamics objective function construction
**Submitting author:** @kbarnhart (<a href="http://orcid.org/0000-0001-5682-455X">Katherine Barnhart</a>)
**Repository:** <a href="https://github.com/TerrainBento/umami" target ="_blank">https://github.com/TerrainBento/umami</a>
**Version:** v1.0.1
