### 2019-10-02 #1776 @tristan-salles
I would like to thank the authors for this contribution, it has been a very enjoyable experience to explore this software and I am pleased to recommend publication, pending the minor issues highlighted below being resolved.

## Version

I noticed that the corresponding version for **umami** is still missing (from `versioneer.py` it is Version: 0.18). It would be good to clarify this, it is especially important for the `pip install` process as mentioned below.

## Installation

I was able to install the code via `conda` and from the `source code` following the instructions but I encountered an issue when using `pip` / `pip3`. It returned the following error message:

```bash
> pip install umami
 Collecting umami
  ERROR: Could not find a version that satisfies the requirement umami (from versions: none)
ERROR: No matching distribution found for umami
```

## Functionality

Working my way through the binder examples illustrate the functionality claims made in the documentation and the paper. More broadly I believe that Umami will be a valuable tool for researchers across the geosciences.

> Potentially the authors can also provide in their readthedocs documentation an additional page on the required steps to use it with other packages than terrainbento and the Landlab Toolkit...

## Documentation

Going through the documentation I was really pleased with the details and quality of what has been produced. I also ❤️ the logo! The dependencies are listed in the `environment-dev.yml` file and are handled via both `conda` and `pip` managers making it really user-friendly. The core functionality (API) is also documented to a good level in Umami readthedocs website.

## Example usage

I went through the four part series of introductory notebooks and was impressed by the quality of the examples (especially Part 3 & 4). I think this is a great tool to calculate objective functions or objective function components for models that run fast.

On Binder I noted that when launching the first notebook, the first cell of the *Part 1: Introduction to Umami and the `Metric` Class* contains a previous output that will need to be cleaned before building the binder container:
```python
<ipython-input-1-6ae86a672d39> in <module>
      4 from io import StringIO
      5 from landlab import RasterModelGrid, imshow_grid
----> 6 from umami import Metric
      7 from umami.calculations import aggregate

ModuleNotFoundError: No module named 'umami'
```

## Paper

The paper clearly outlines the motivation for the development of this package. It reads clearly and follows JOSS guidelines.

In conclusion, I am delighted to have had the opportunity to review this paper, and have been happy with the quality of the software, documentation and examples. I look forward to use this package with my own software!
