### 2019-10-02 TerrainBento/umami#5 @kbarnhart
pip installation not functioning
As pointed out by @tristan-salles in [his review of the JOSS submission](https://github.com/openjournals/joss-reviews/issues/1776#issuecomment-537319176) the pip installation currently does not work.

He reports the following error:

```bash
> pip install umami
 Collecting umami
  ERROR: Could not find a version that satisfies the requirement umami (from versions: none)
ERROR: No matching distribution found for umami
```
