### 2019-10-02 TerrainBento/umami#6 @kbarnhart
Make version 1.0 release for JOSS
@tristan-salles points out on his [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1776#issuecomment-537319176) that the current version is not the 1.0 version referenced in the joss submission.

This is because I wanted to ensure that any changes necessary to the code base/docs/etc that might happen in response to the review would happen before the 1.0 version release.

This PR exists as a reminder to make this release.
