### 2019-10-02 TerrainBento/umami#7 @kbarnhart
Clean up the output of notebook cells in examples
As discussed in the [joss review by @tristan-salles](https://github.com/openjournals/joss-reviews/issues/1776#issuecomment-537319176) the first notebook (*Part 1: Introduction to Umami and the Metric*) has some previous output that needs to be removed.
