### 2019-10-02 TerrainBento/umami#8 @kbarnhart
Add discussion to docs of using umami with packages that are not terrainbento or Landlab. 
In his [joss review](https://github.com/openjournals/joss-reviews/issues/1776) @tristian-salles recommends

> Potentially the authors can also provide in their readthedocs documentation an additional page on the required steps to use it with other packages than terrainbento and the Landlab Toolkit...

Will do.
