### 2019-10-02 #1776 @kbarnhart
@tristan-salles thanks for your review. Based on it I made four Issues in the repo and will address them. Doing fieldwork tomorrow, so will be later this week/early next week.
