### 2019-10-02 TerrainBento/umami#5 @kbarnhart
@tristan-salles: would you be able to provide me with a few more details of your compute environment and/or installation procedure so that I can address this issue? I just tried to reproduce it on my computer (OSX, Anaconda Python v3.7) and was not able to.

I tried installing umami in my base conda environment and in a new conda environment. Both worked for me.

```bash
$ conda create --name test_umami
$ conda activate test_umami
$ conda install pip
$ pip install umami
```
I was then able to start python and import umami.

Thanks again for your review!
