### 2019-10-03 TerrainBento/umami#5 @tristan-salles
Hi @kbarnhart I did try using a linux ubuntu installation not running conda... this might be the issue.
Looking at the installed packages on my current environment I have the following list:

```python
import pip
installed_packages = pip.get_installed_distributions()
installed_packages_list = sorted(["%s==%s" % (i.key, i.version)
     for i in installed_packages])
print(installed_packages_list)
```

```bash
['affine==2.2.2', 'aiohttp==3.5.4', 'alabaster==0.7.12', 'appdirs==1.4.3', 'async-timeout==3.0.1', 'attrs==19.1.0', 'babel==2.7.0', 'backcall==0.1.0', 'bleach==3.1.0', 'cartopy==0.17.0', 'certifi==2019.6.16', 'cftime==1.0.3.4', 'chardet==3.0.4', 'click-plugins==1.1.1', 'click==7.0', 'cligj==0.5.0', 'cycler==0.10.0', 'cython==0.29.13', 'decorator==4.4.0', 'defusedxml==0.6.0', 'descartes==1.1.0', 'docutils==0.14', 'easyprocess==0.2.7', 'entrypoints==0.3', 'escape==0.1', 'fastfunc==0.1.7', 'fillit==0.0.1', 'fillpy==0.0.1', 'geographiclib==1.49', 'geopy==1.20.0', 'glem==0.1', 'h5py==2.9.0', 'idna-ssl==1.1.0', 'idna==2.8', 'imagesize==1.1.0', 'ipykernel==5.1.1', 'ipyparallel==6.2.4', 'ipython-genutils==0.2.0', 'ipython==7.6.1', 'ipywidgets==7.5.0', 'jedi==0.14.0', 'jinja2==2.10.1', 'jsonschema==3.0.1', 'jupyter-client==5.2.4', 'jupyter-console==6.0.0', 'jupyter-core==4.5.0', 'jupyter-server-proxy==1.1.0', 'jupyter==1.0.0', 'kiwisolver==1.1.0', 'lavavu==1.4.0', 'markupsafe==1.1.1', 'matplotlib==3.0.3', 'meshio==2.3.9', 'meshplex==0.11.6', 'mistune==0.8.4', 'mock==3.0.5', 'mpi4py==3.0.2', 'multidict==4.5.2', 'nbconvert==5.5.0', 'nbformat==4.4.0', 'netcdf4==1.5.2', 'notebook==5.7.8', 'numpy==1.17.0', 'packaging==19.0', 'pandas==0.24.2', 'pandocfilters==1.4.2', 'parso==0.5.0', 'pathlib==1.0.1', 'petsc4py==3.11.0', 'pexpect==4.7.0', 'pickleshare==0.7.5', 'pillow==6.1.0', 'pint==0.9', 'pipdate==0.3.2', 'plotly==3.10.0', 'pockets==0.7.2', 'prometheus-client==0.7.1', 'prompt-toolkit==2.0.9', 'ptyprocess==0.6.0', 'pybind11==2.3.0', 'pycpt==0.2.0', 'pygeotools==0.6.0', 'pygments==2.4.2', 'pygmsh==5.0.1', 'pykdtree==1.3.1', 'pyparsing==2.4.0', 'pyrsistent==0.15.2', 'pyshp==2.1.0', 'python-dateutil==2.8.0', 'pytz==2019.1', 'pyvirtualdisplay==0.2.4', 'pyvtk==0.5.18', 'pyzmq==18.0.2', 'qtconsole==4.5.1', 'rasterio==1.0.24', 'requests==2.22.0', 'retrying==1.3.3', 'ruamel.yaml==0.15.98', 'scipy==1.3.1', 'scons==3.0.5', 'seaborn==0.9.0', 'send2trash==1.5.0', 'setuptools==41.0.1', 'shapely==1.6.4.post2', 'simpervisor==0.3', 'six==1.12.0', 'snowballstemmer==1.9.0', 'snuggs==1.4.6', 'sphinx-rtd-theme==0.4.3', 'sphinx==2.1.2', 'sphinxcontrib-applehelp==1.0.1', 'sphinxcontrib-devhelp==1.0.1', 'sphinxcontrib-htmlhelp==1.0.2', 'sphinxcontrib-inlinesyntaxhighlight==0.2', 'sphinxcontrib-jsmath==1.0.1', 'sphinxcontrib-napoleon==0.7', 'sphinxcontrib-qthelp==1.0.2', 'sphinxcontrib-serializinghtml==1.1.3', 'sphinxemoji==0.1.4', 'stripy==1.0.2', 'tabulate==0.8.3', 'terminado==0.8.2', 'testpath==0.4.2', 'tornado==6.0.3', 'traitlets==4.3.2', 'typing-extensions==3.7.4', 'urllib3==1.25.3', 'vtk==8.1.2', 'wcwidth==0.1.7', 'webencodings==0.5.1', 'wheel==0.33.4', 'widgetsnbextension==3.5.0', 'yarl==1.3.0']
```

After googling a bit for the error message I did find on [stackoverflow](https://stackoverflow.com/questions/32302379/could-not-find-a-version-that-satisfies-the-requirement-package) a potential work around that actually didn't help in my case...

```python3
python3 -m pip install <pkg>
```

If I am the only one having the issue it must be related to my specific machine...
