### 2019-10-06 TerrainBento/umami#5 @kbarnhart
@tristan-salles thanks for the additional information, very helpful.

@mcflugen do you have any recommendations regarding this issue? And/or do you have any thoughts on how we might test installs on systems that have pip only, and no conda install?

(also @mcflugen can you accept my invite to become a contributor to this repo).
