### 2019-10-06 TerrainBento/umami#5 @mcflugen
@kbarnhart I think the problem is that you've built a wheel and uploaded it to PyPI. The version that's on PyPI was built for a specific version of linux. If you do want to distribute a binary package for linux, it should be labeled as *manylinux*. You can have a look at *landlab* for an example of how to do this.

However, because *umami* is pure Python, you should be able to get away with uploading a source distribution, which will then be available for all platforms. You can do this with `python setup.py sdist`.
