### 2019-10-06 TerrainBento/umami#5 @kbarnhart
@mcflugen quick question...

I thought that the fact that the wheel I uploaded was a `none-any` wheel meant that it was good for all Linux-based (including OSX) and for Windows. I don't understand why the current wheel works for OSX, some linux, and windows, but not all linux. And thus I don't understand why switching from a `none-any` wheel to a source distribution will fix things. Do you understand why this might be?

(sidenote, just tested `pip install umami` on a Windows machine I have access to and my Linux box running Ubuntu 18.04 LTS... both installed OK).

A final note: I tried copying the Landlab _manylinux_ approach when I originally set up the CI for umami, but because umami is pure python all the wheel's had the same name (and so all but the first deploy stages failed due to an upload error from pypi).
