### 2019-10-06 TerrainBento/umami#5 @kbarnhart
@tristan-salles can I ask you to try installing with

```bash
pip install umami -i https://pypi.org/simple
```

That is, explicitly stating to [`pip install`](https://pip.pypa.io/en/stable/reference/pip_install/#cmdoption-index-url) which index-url to get umami from?
