### 2019-10-06 TerrainBento/umami#5 @mcflugen
@tristan-salles Could you also run the following?
```bash
$ pip debug
```
and then, possibly,
```bash
$ pip debug -v
```
This should tell us which tags your *pip* is looking for and if `py3-none-any`, which is currently what's on PyPI, is one of those tags.
