### 2019-10-06 TerrainBento/umami#5 @mcflugen
@kbarnhart I don't know why his *pip* can't find the distribution when most others can. Maybe it's an old version of *pip*?
