### 2019-10-06 TerrainBento/umami#5 @tristan-salles
Hi @kbarnhart, this gives me the following...

```bash
root@300509464e10:/live/share/global/globalData# pip install umami -i https://pypi.org/simple
Collecting umami
ERROR: Could not find a version that satisfies the requirement umami (from versions: none)
ERROR: No matching distribution found for umami
```
