### 2019-10-07 TerrainBento/umami#5 @tristan-salles
Hi @mcflugen,

when I ran the `pip debug` I get:
```bash
root@300509464e10:/live/share/global/globalData# pip debug
WARNING: This command is only meant for debugging. Do not use this with automation for parsing and getting these details, since the output and options of this command may change without notice.
pip version: pip 19.2.3 from /usr/local/lib/python3.5/dist-packages/pip (python 3.5)
sys.version: 3.5.3 (default, Sep 27 2018, 17:25:39)
[GCC 6.3.0 20170516]
sys.executable: /usr/bin/python3
sys.getdefaultencoding: utf-8
sys.getfilesystemencoding: utf-8
locale.getpreferredencoding: UTF-8
sys.platform: linux
sys.implementation:
  name: cpython
Compatible tags: 30
  cp35-cp35m-manylinux2010_x86_64
  cp35-cp35m-manylinux1_x86_64
  cp35-cp35m-linux_x86_64
  cp35-abi3-manylinux2010_x86_64
  cp35-abi3-manylinux1_x86_64
  cp35-abi3-linux_x86_64
  cp35-none-manylinux2010_x86_64
  cp35-none-manylinux1_x86_64
  cp35-none-linux_x86_64
  cp34-abi3-manylinux2010_x86_64
  ...
  [First 10 tags shown. Pass --verbose to show all.]
```

and `pip debug -v`

```bash
root@300509464e10:/live/share/global/globalData# pip debug -v
WARNING: This command is only meant for debugging. Do not use this with automation for parsing and getting these details, since the output and options of this command may change without notice.
pip version: pip 19.2.3 from /usr/local/lib/python3.5/dist-packages/pip (python 3.5)
sys.version: 3.5.3 (default, Sep 27 2018, 17:25:39)
[GCC 6.3.0 20170516]
sys.executable: /usr/bin/python3
sys.getdefaultencoding: utf-8
sys.getfilesystemencoding: utf-8
locale.getpreferredencoding: UTF-8
sys.platform: linux
sys.implementation:
  name: cpython
Compatible tags: 30
  cp35-cp35m-manylinux2010_x86_64
  cp35-cp35m-manylinux1_x86_64
  cp35-cp35m-linux_x86_64
  cp35-abi3-manylinux2010_x86_64
  cp35-abi3-manylinux1_x86_64
  cp35-abi3-linux_x86_64
  cp35-none-manylinux2010_x86_64
  cp35-none-manylinux1_x86_64
  cp35-none-linux_x86_64
  cp34-abi3-manylinux2010_x86_64
  cp34-abi3-manylinux1_x86_64
  cp34-abi3-linux_x86_64
  cp33-abi3-manylinux2010_x86_64
  cp33-abi3-manylinux1_x86_64
  cp33-abi3-linux_x86_64
  cp32-abi3-manylinux2010_x86_64
  cp32-abi3-manylinux1_x86_64
  cp32-abi3-linux_x86_64
  py3-none-manylinux2010_x86_64
  py3-none-manylinux1_x86_64
  py3-none-linux_x86_64
  cp35-none-any
  cp3-none-any
  py35-none-any
  py3-none-any
  py34-none-any
  py33-none-any
  py32-none-any
  py31-none-any
  py30-none-any
```

This is returning the `py3-none-any` tag...

Sorry for the pain!
