### 2019-10-07 TerrainBento/umami#5 @mcflugen
@tristan-salles Thank you!

I think the problem is that you have Python 3.5 installed (is that right? From `pip debug` it appears so). If I do the following, I can reproduce your problem on my Mac,
```
$ conda create -n py35 python=3.5
$ conda activate py35
$ pip install umami
Collecting umami
  Could not find a version that satisfies the requirement umami (from versions: )
No matching distribution found for umami
```

However, if I try to install *umami* within a Python 3.7 environment,
```bash
$ conda create -n py37 python=3.7
$ conda activate py37
$ pip install umami
```
*umami* installs just fine.
