### 2019-10-07 TerrainBento/umami#5 @mcflugen
@kbarnhart I think the Python version requirement is coming from the `python_requires` keyword in the `setup.py` file.
