### 2019-10-07 TerrainBento/umami#5 @kbarnhart
@tristan-salles thanks for your help! And @mcflugen thanks for figuring it out! I'll fix the `setup.py` and make a new release.
