### 2019-10-07 TerrainBento/umami#5 @kbarnhart
@mcflugen As part of making and testing this new release I've run into an issue that I need your help with.

As Landlab only puts a py3.6 and py3.7 release on PyPI, this means that `pip install umami` fails for python versions 3.6 and lower. After making the current v0.1.3 release for umami I tested an install with py3.4 and py3.5 (I tried to make a py3.0, py3.1, py3.2, and py3.3 but conda didn't support it).

On both the py3.4 and py3.5 tests I ran into the same problem:

```bash
$ conda create -n py34 python=3.4
$ conda activate py34
$ pip install umami
Collecting umami
  Cache entry deserialization failed, entry ignored
  Downloading https://files.pythonhosted.org/packages/38/b7/157fb76582aca42c052f5833fdc33b1fc851c34724c
645a88c08c2d00cd7/umami-0.1.3-py3-none-any.whl
Collecting numpy (from umami)
  Cache entry deserialization failed, entry ignored
  Downloading https://files.pythonhosted.org/packages/db/ec/93ddd4696e9cce0ffb8429516a8ba0d0ee95911cbbadde2d23665b62ad39/numpy-1.16.5.zip (5.1MB)
    100% |████████████████████████████████| 5.1MB 265kB/s
Collecting landlab>=1.10.1 (from umami)
  Cache entry deserialization failed, entry ignored
  Could not find a version that satisfies the requirement landlab>=1.10.1 (from umami) (from
 versions: 0.1.7.macosx-10.5-x86_64, 0.1.6, 0.1.7, 0.1.8, 0.1.9, 0.1.12, 0.1.14, 0.1.15, 0.1.18, 0.1.21,
 0.1.22, 0.1.23, 0.1.24, 0.1.25, 0.1.26, 0.1.27, 0.1.28, 0.1.29, 0.1.33, 0.1.35, 0.1.37, 0.1.38, 0.1.39, 0.1.40,
 0.1.41, 0.2.0, 0.2.1, 0.2.2, 0.2.3, 0.5.0, 1.0.0a0, 1.0.0b0, 1.0.0b1, 1.0.0b3, 1.0.0b4, 1.0.0b6, 1.0.0b8,
 1.0.0b10, 1.0.0b11, 1.0.0b12, 1.0.0, 1.0.1, 1.0.2, 1.0.3, 1.1.0, 1.2.0, 1.3.0, 1.3.1, 1.4.0, 1.5.0, 1.5.1, 1.5.2,
 1.5.3, 1.5.4, 1.9.0b0, 1.9.0b1, 1.9.0b3, 1.9.0b4, 1.9.0b5, 1.9.0)
No matching distribution found for landlab>=1.10.1 (from umami)
You are using pip version 9.0.1, however version 19.2.3 is available.
You should consider upgrading via the 'pip install --upgrade pip' command.
```

My question: Is this a problem? In theory a user working on py3.5 could download landlab from source and compile, and then get umami from PyPI...

What do you think makes the most sense?
