### 2019-10-07 TerrainBento/umami#5 @kbarnhart
Also, thanks @tristan-salles for helping us identify and diagnose this problem. I'm really glad you found it. 🎉
