### 2019-10-07 TerrainBento/umami#5 @tristan-salles
Hi @kbarnhart & @mcflugen I did try with python 3.8 and the pip package is working as expected...
