### 2019-10-08 TerrainBento/umami#5 @mcflugen
@kbarnhart I think relaxing the `python_requires` to `>=3` for *umami* is good and probably all that needs to be done. If a user really wants to use an old version of Python they can, as you say, install *landlab* (and any other packages that aren't available for their version of Python) from source.

We can certainly discuss building *landlab* for old Python versions but I'm not certain it would be worth it - particularly since one can always create a separate environment for their install.

Its too bad the error message that *pip* gives isn't more informative.
