### 2019-10-08 TerrainBento/umami#5 @kbarnhart
Thanks for that input @mcflugen. I'm going to close this issue now.
