### 2019-10-08 TerrainBento/umami#7 @kbarnhart
I cleaned this up in #9 and verified that the old output is no longer there by opening each notebook on Binder.
