### 2019-10-09 17ceef7
fix: add rasterio dependency for notebooks
