+++ environment-dev.yml
@@ -20,3 +20,4 @@ dependencies:
    - tqdm
    - plotnine>=0.6.0
    - terrainbento>=1.1
+   - rasterio
