+++ appveyor.yml
@@ -28,7 +28,7 @@ install:
   - cmd: set PYTHONUNBUFFERED=1
   - cmd: conda config --set always_yes yes
   - cmd: pip install pytest
-  - cmd: pip install jupyter pandas plotnine holoviews terrainbento tqdm
+  - cmd: pip install jupyter pandas plotnine holoviews terrainbento tqdm rasterio
   - cmd: conda info
   - cmd: conda list
 
