### 2019-10-09 fca8229
add gdal dependency
