+++ environment-dev.yml
@@ -20,4 +20,5 @@ dependencies:
    - tqdm
    - plotnine>=0.6.0
    - terrainbento>=1.1
+   - gdal
    - rasterio
