### 2019-10-09 a182d0b
travis was successful without gdal. so leaving as was
