### 2019-10-09 1fb8dd2
trying a pip install gdal before other packages.
