+++ appveyor.yml
@@ -28,9 +28,7 @@ install:
   - cmd: set PYTHONUNBUFFERED=1
   - cmd: conda config --set always_yes yes
   - cmd: pip install pytest
-  - cmd: conda config --add channels conda-forge
-  - cmd: conda config --set channel_priority strict
-  - cmd: conda install gdal
+  - cmd: pip install GDAL
   - cmd: pip install jupyter pandas plotnine holoviews terrainbento tqdm rasterio
   - cmd: conda info
   - cmd: conda list
