### 2019-10-09 54e0deb
use conda for rasterio on appveyor
