+++ appveyor.yml
@@ -28,8 +28,8 @@ install:
   - cmd: set PYTHONUNBUFFERED=1
   - cmd: conda config --set always_yes yes
   - cmd: pip install pytest
-  - cmd: pip install GDAL
-  - cmd: pip install jupyter pandas plotnine holoviews terrainbento tqdm rasterio
+  - cmd: pip install jupyter pandas plotnine holoviews terrainbento tqdm
+  - cmd: conda install rasterio -c conda-forge
   - cmd: conda info
   - cmd: conda list
 
