### 2019-10-10 e7ed181
get landlab from conda-forge instead of pip
