### 2019-10-11 #1776 @sgrieve
Firstly I would like to thank the author for this contribution. I have been very impressed by the quality of the software presented here, as well as it's potential utility to a wide range of users across disciplines. I am happy to recommend publication, pending the resolution of the **minor** issues outlined below.

I would like to also note that the way this software has been developed, documented, tested and supported is of an exceptionally high standard. I would be happy to use this codebase as an exemplar of good practice in my teaching.

# Installation

I had no problems installing `umami` using pip on MacOS 10.14.6, inside a clean virtualenv running python 3.6.3. The install docs were clear, all dependences are listed in the environment file, with comments explaining where they are used.

# Functionality

The functionality of the software is clearly outlined in the series of notebooks, and these provide an excellent starting point for anyone wanting to use `umami` in their research. All of the claims of the functionality of the software are valid, and supported by clear documentation and examples.

# Documentation

The documentation is of a very high standard, with very detailed API docs, higher level documentation on the readthedocs site and detailed jupyter notebooks outlining the usage of the software in a range of scenarios.

# Community guidelines

There is a clear invitation to contribute to the software, although no detail on the preferred method of or type of contributions is given (eg opening issues, editing docs, PRs, etc). The code of conduct is excellent and I will be learning from this example of good practice for my next project.

# Tests

It is clear that the tests have been diligently designed to identify potential errors in the code. CI is set up on both Travis and Appveyor, ensuring the widest possible range of environments are being tested. Reading the tests, they are well written, and self-documenting, with good use of fixtures and other pytest and numpy features.

# Paper

The paper is excellently written and sets out the general problem being addressed and the wider context of the discipline. The second section of the paper neatly presents the software, its functionality, and its novelty. The description of the multiple output formats, and crucially, the rationale behind the choice of formats is great.

I noticed that one of the in text references is inconsistent with the others, with `(B. M. Adams et al., 2019)` including initials where other in text references only use surnames.
