### 2019-10-11 TerrainBento/umami#12 @kbarnhart
Improve contribution guidelines
In a [JOSS review](https://github.com/openjournals/joss-reviews/issues/1776#issuecomment-541069220) @sgrieve points out:

> There is a clear invitation to contribute to the software, although no detail on the preferred method of or type of contributions is given (eg opening issues, editing docs, PRs, etc).

(@kbarnhart  will improve the contribution guidelines to provide more detail regarding the preferred method of contribution and the type of contributions that are welcome).
