### 2019-10-11 TerrainBento/umami#8 @kbarnhart
@tristian-salles, I made a new tutorial to address this issue [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/TerrainBento/umami/master?filepath=notebooks%2FOtherIO_options.ipynb)

If you have any comments on the tutorial, I'd welcome incorporating them before I close the issue.

Thanks!
