### 2019-10-11 7ea626f
Add more specific detail to describe contributions.
