+++ README.md
@@ -97,7 +97,18 @@ any reason, please make
 
 # Contribute to umami
 
-All contributions are welcome.
+All contributions are welcom and appreciated. Feel free to:
+
+- Make [an issue](https://github.com/TerrainBento/umami/issues) to ask a
+  question. Your question will help others in the future.
+- Make [an issue](https://github.com/TerrainBento/umami/issues) to report a
+  bug or a potential improvement. We will work to fix it. If you have an idea
+  about how to fix it, please feel free to propose it, or make a Pull Request.  
+- [Fork the repository](https://help.github.com/en/articles/fork-a-repo), make
+  changes to the source code on a
+  [development branch](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow),
+  and submit a Pull Request to have your changes brought into the umami
+  repository. **No contribution to the code base or documentation is too small.**
 
 Contributors and maintainers to this project are are expected to abide the [Contributor Code of Conduct](https://github.com/TerrainBento/umami/blob/master/CODE_OF_CONDUCT.md).
 
