+++ joss/paper.bib
@@ -23,7 +23,7 @@
     }
 
 @article{adams2019dakota,
-    author = {Adams and Eldred and Geraci and Hooper and Jakeman and Maupin and Monschk and Rushdi and Stephens and Swiler and Wildey and Bohnhoff and Dalbey and Ebeida and Eddy and Hough and Khalil and Hu and Ridgway and Vigil and Winokur},
+    author = {Adams, Brian M and Eldred, Michael S and Geraci, Gianluca and Hooper, Russell W and Jakeman, John D and Maupin, Kathryn A and Monschk, Jason A and Rushdi, Ahmad A and Stephens, J Adam and Swiler, Laura P and Wildey, Timothy M and Bohnhoff, William J and Dalbey, Keith R and Ebeida, Mohamed S and Eddy, John P and Hough, Patricia D and Khalil, Mohammad and Hu, Kenneth T and Ridgway, Elliott M and Vigil, Dena M and Winokur, Justin G},
     title = {{Dakota, A Multilevel Parallel Object-Oriented Framework for Design Optimization, Parameter Estimation, Uncertainty Quantification, and Sensitivity Analysis: Version 6.10 User{\textquoteright}s Manual}},
     journal = {Sandia National Laboratories, Tech. Rep. SAND2014-4633},
     year = {2019}
