+++ joss/paper.bib
@@ -23,7 +23,7 @@
     }
 
 @article{adams2019dakota,
-    author = {Adams, Brian M and Eldred, Michael S and Geraci, Gianluca and Hooper, Russell W and Jakeman, John D and Maupin, Kathryn A and Monschk, Jason A and Rushdi, Ahmad A and Stephens, J Adam and Swiler, Laura P and Wildey, Timothy M and Bohnhoff, William J and Dalbey, Keith R and Ebeida, Mohamed S and Eddy, John P and Hough, Patricia D and Khalil, Mohammad and Hu, Kenneth T and Ridgway, Elliott M and Vigil, Dena M and Winokur, Justin G},
+    author = {Adams, B. M. and Eldred, S. and Geraci, G. and Hooper, R. W. and Jakeman, J. D. and Maupin, K. A. and Monschk, J. A. and Rushdi, A. A. and Stephens, J. A. and Swiler, L. P. and Wildey, T. M. and Bohnhoff, W. J. and Dalbey, K. R. and Ebeida, M. S. and Eddy, J. P. and Hough, P. D. and Khalil, M. and Hu, K. T. and Ridgway, E. M. and Vigil, D. M. and Winokur, J. G.},
     title = {{Dakota, A Multilevel Parallel Object-Oriented Framework for Design Optimization, Parameter Estimation, Uncertainty Quantification, and Sensitivity Analysis: Version 6.10 User{\textquoteright}s Manual}},
     journal = {Sandia National Laboratories, Tech. Rep. SAND2014-4633},
     year = {2019}
