### 2019-10-11 TerrainBento/umami#14 @kbarnhart
Fix bibliography
@sgrieve pointed out in the [Joss review](https://github.com/openjournals/joss-reviews/issues/1776#issuecomment-541069220) that one of the in-text references was inconsistent with the others. Specifically

> I noticed that one of the in text references is inconsistent with the others, with (B. M. Adams et al., 2019) including initials where other in text references only use surnames.
