### 2019-10-11 bfc0236
bit of .bib clean up for capitalization.
