+++ joss/paper.bib
@@ -1,44 +1,41 @@
 @Article{hobley2017creative,
     AUTHOR = {Hobley, D. E. J. and Adams, J. M. and Nudurupati, S. S. and Hutton, E. W. H. and Gasparini, N. M. and Istanbulluoglu, E. and Tucker, G. E.},
-    TITLE = {Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics},
+    TITLE = {{Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics}},
     JOURNAL = {Earth Surface Dynamics},
     VOLUME = {5},
     YEAR = {2017},
     NUMBER = {1},
     PAGES = {21--46},
-    URL = {https://www.earth-surf-dynam.net/5/21/2017/},
     DOI = {10.5194/esurf-5-21-2017}
     }
 
 @Article{barnhart2019terrain,
     AUTHOR = {Barnhart, K. R. and Glade, R. C. and Shobe, C. M. and Tucker, G. E.},
-    TITLE = {Terrainbento 1.0: a Python package for multi-model analysis in long-term drainage basin evolution},
+    TITLE = {{Terrainbento 1.0: a Python package for multi-model analysis in long-term drainage basin evolution}},
     JOURNAL = {Geoscientific Model Development},
     VOLUME = {12},
     YEAR = {2019},
     NUMBER = {4},
     PAGES = {1267--1297},
-    URL = {https://www.geosci-model-dev.net/12/1267/2019/},
     DOI = {10.5194/gmd-12-1267-2019}
     }
 
 @article{adams2019dakota,
     author = {Adams, B. M. and Eldred, S. and Geraci, G. and Hooper, R. W. and Jakeman, J. D. and Maupin, K. A. and Monschk, J. A. and Rushdi, A. A. and Stephens, J. A. and Swiler, L. P. and Wildey, T. M. and Bohnhoff, W. J. and Dalbey, K. R. and Ebeida, M. S. and Eddy, J. P. and Hough, P. D. and Khalil, M. and Hu, K. T. and Ridgway, E. M. and Vigil, D. M. and Winokur, J. G.},
     title = {{Dakota, A Multilevel Parallel Object-Oriented Framework for Design Optimization, Parameter Estimation, Uncertainty Quantification, and Sensitivity Analysis: Version 6.10 User{\textquoteright}s Manual}},
-    journal = {Sandia National Laboratories, Tech. Rep. SAND2014-4633},
+    journal = {{Sandia National Laboratories, Tech. Rep. SAND2014-4633}},
     year = {2019}
     }
 
 
 @Article{forte2019tak,
     AUTHOR = {Forte, A. M. and Whipple, K. X.},
-    TITLE = {Short communication: The Topographic Analysis Kit (TAK) for TopoToolbox},
+    TITLE = {{Short communication: The Topographic Analysis Kit (TAK) for TopoToolbox}},
     JOURNAL = {Earth Surface Dynamics},
     VOLUME = {7},
     YEAR = {2019},
     NUMBER = {1},
     PAGES = {87--95},
-    URL = {https://www.earth-surf-dynam.net/7/87/2019/},
     DOI = {10.5194/esurf-7-87-2019}
     }
 
@@ -50,7 +47,6 @@
     YEAR = {2017},
     NUMBER = {3},
     PAGES = {369--385},
-    URL = {https://www.earth-surf-dynam.net/5/369/2017/},
     DOI = {10.5194/esurf-5-369-2017}
     }
 
@@ -61,42 +57,38 @@
     volume = {119},
     number = {2},
     pages = {138-152},
[…]
