### 2019-10-11 TerrainBento/umami#14 @kbarnhart
@arfon, I think you may be the correct person to ask for help with this issue.

I have a bibliography entry that looks like:

```
@article{adams2019dakota,
    author = {Adams, B. M. and Eldred, S. and Geraci, G. and Hooper, R. W. and
Jakeman, J. D. and Maupin, K. A. and Monschk, J. A. and Rushdi, A. A. and
Stephens, J. A. and Swiler, L. P. and Wildey, T. M. and Bohnhoff, W. J. and
Dalbey, K. R. and Ebeida, M. S. and Eddy, J. P. and Hough, P. D. and Khalil, M.
and Hu, K. T. and Ridgway, E. M. and Vigil, D. M. and Winokur, J. G.},
    title = {{Dakota, A Multilevel Parallel Object-Oriented Framework for Design
Optimization, Parameter Estimation, Uncertainty Quantification, and Sensitivity
Analysis: Version 6.10 User{\textquoteright}s Manual}},
    journal = {{Sandia National Laboratories, Tech. Rep. SAND2014-4633}},
    year = {2019}
    }
```

I need it to make an in-text reference that looks like:

Adams et al., 2019

Instead I'm getting

B. M. Adams et al., 2019

Previously, when I used a 2017 version of this bib entry AND referenced another paper that has a J. N. Adams et al., 2017 reference, I ran into issues regarding the citation because of bib latex's rules regarding disambiguation. But I'm not citing that paper in this submission (though J. N. Adams is a second author on one of the papers).

I just looked at what I **think** whedon might be using to generate the paper [Latex Template](https://github.com/openjournals/whedon/blob/master/resources/latex.template), and haven't found anything that has helped me figure out why this reference is formatting in this way.

Do you have any recommendations?
