### 2019-10-11 #1776 @arfon
@kbarnhart - I've upgraded Pandoc (from `v2.0.5` -> `v2.7.3`) which seems to have fixed the issue with the `Adams et al.` citation string / cc https://github.com/TerrainBento/umami/issues/14#issuecomment-541239203
