### 2019-10-11 TerrainBento/umami#14 @arfon
@kbarnhart - not sure what was going on here but the version of Pandoc Whedon was using was pretty old so I've updated it and that seems to have fixed things `¯\_(ツ)_/¯`
