### 2019-10-14 #1776 @meg-simula
@sgrieve and @tristan-salles Thanks for completing the reviews, much appreciated.
@kbarnhart Looks like you are well underway handling the feedback. Let me know if you have further questions or when you are ready to proceed further.
