### 2019-10-14 #1776 @kbarnhart
Sounds good @meg-simula, I'm going to give @tristan-salles a little more time to see if he has any further recommendations to address [Issue #8](https://github.com/TerrainBento/umami/issues/8). Once I close that, I'll make the 1.0 release (which will address the [last remaining issue](https://github.com/TerrainBento/umami/issues/6)), and I'll let you know that I'm ready to proceed.

Thanks all for constructive comments!
