### 2019-10-14 TerrainBento/umami#8 @kbarnhart
Thanks @tristan-salles. Closing.
