### 2019-10-14 #1776 @tristan-salles
Hi @meg-simula & @kbarnhart, I have just reviewed the issue #8 and it looks really good and I am happy to recommend this publication.
