### 2019-10-15 #1776 @kbarnhart
@meg-simula, I've now resolved all the issues related to the reviews and made the v1.0.0 release.

Let me know what you need from me to move forward. Thanks!
