### 2019-10-16 #1776 @sgrieve
Yes, I wholeheartedly recommend acceptance. 🚀
