### 2019-10-22 #1776 @kbarnhart
thanks for your feedback @sgrieve and @tristan-salles and to @meg-simula for handling this! 🎉

the DOI for the archive on Zenodo is https://doi.org/10.5281/zenodo.3489794
