### 2019-10-28 #1776 @meg-simula
@kbarnhart Could you please do an additional spell-check of the paper? (I assume "hight of mountain ranges" should be "height of ..." e.g.)
