### 2019-10-29 #1776 @kbarnhart
@meg-simula
- [x] I spell checked the document, fixed the typo you noticed, and found one other.
- [x] Incremented the version number.
- [x] Made a Zenodo archive with the same title and author order as the JOSS manuscript (rather than the auto-created Zenodo archive that comes from linking Zenodo with Github).

The new version is: v1.0.1
The new Zenodo archive is: https://doi.org/10.5281/zenodo.3521085

Thank you!
