### 2019-10-29 #1776 @meg-simula
@openjournals/joss-eics This paper is ready to go.
