### 2019-10-29 #1776 @danielskatz
👋 @kbarnhart - please merge the few minor changes in https://github.com/TerrainBento/umami/pull/19
